collector/requirements.txt
google-cloud-storage>=2.0.0
requests>=2.0.0
pandas>=1.0.0
numpy>=1.18.0
google-auth>=2.0.0